// generated/src/app/api-models/premium-master.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class PremiumMaster {
  seqPremId: number;
  seqGroupId: number;
  recordType: string;
  planRiderCode: string;
  effectiveDate: Date;
  endDate: Date;
  benefitPackageId: string;
  lineOfBusiness: string;
  companyCode: string;
  glRefCode: string;
  tieringMethod: string;
  rate1: number;
  rate2: number;
  rate3: number;
  rate4: number;
  rate5: number;
  priceRule1: string;
  priceRule2: string;
  priceSchedule1: string;
  priceSchedule2: string;
  pctOfBilled: number;
  priceOrAdjudicate: string;
  pctAllowed: number;
  matrixDef: string;
  pctOfMatrix: number;
  adminFee: number;
  adFeeCompCode: string;
  adFeeGlCode: string;
  specPgmProcessCd: string;
  geographicRegion: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  adFeeLevelCode: string;
  exclusionType: string;
  exclusionPeriod: number;
  applyPcp: string;
  reasonCode: string;
  cobCalcMethod: string;
  commissionMatrixDef: string;
  commissionCompanyCode: string;
  commissionGlRefCode: string;
  rateCategory: string;
  instCobCalcMethod: string;
  depDetermRuleCode: string;
  cobPolicyFlag: string;
  cobExcludeCapClaimCode: string;
  rateModifyDate: Date;
  claimFilingIndicator: string;
  dentalCobCalcMethod: string;
  excludeCapClmDtl: string;
}